Provide file_operations::unlocked_ioctl.
file_operations::ioctl is no longer supported in recent kernels.
compat_ioctl is not running under the BKL, so the code has to do its own
locking anyway (although not everything currently does).

+++ ncr-int.h
@@ -104,7 +104,7 @@ struct ncr_lists {
 void* ncr_init_lists(void);
 void ncr_deinit_lists(struct ncr_lists *lst);
 
-int ncr_ioctl(struct ncr_lists *lst, unsigned int cmd, unsigned long arg);
+long ncr_ioctl(struct ncr_lists *lst, unsigned int cmd, unsigned long arg);
 
 /* key derivation */
 int ncr_key_derive(struct ncr_lists *lst, void __user* arg);
